@@ -11,18 +11,18 @@
 
 namespace Sylius\Bundle\CartBundle\Twig;
 
+use Sylius\Bundle\CartBundle\Model\CartInterface;
 use Sylius\Bundle\CartBundle\Provider\CartProviderInterface;
 use Sylius\Bundle\ResourceBundle\Model\RepositoryInterface;
-use Symfony\Component\Form\FormFactory;
-use Twig_Extension;
-use Twig_Function_Method;
+use Symfony\Component\Form\FormFactoryInterface;
+use Symfony\Component\Form\FormView;
 
 /**
  * Sylius cart engine twig extension.
  *
  * @author Paweł Jędrzejewski <pjedrzejewski@diweb.pl>
  */
-class SyliusCartExtension extends Twig_Extension
+class SyliusCartExtension extends \Twig_Extension
 {
     /**
      * Cart provider.
@@ -41,7 +41,7 @@ class SyliusCartExtension extends Twig_Extension
     /**
      * Form factory.
      *
-     * @var FormFactory
+     * @var FormFactoryInterface
      */
     protected $formFactory;
 
@@ -50,9 +50,9 @@ class SyliusCartExtension extends Twig_Extension
      *
      * @param CartProviderInterface $cartProvider
      * @param RepositoryInterface   $cartItemRepository
-     * @param FormFactory           $formFactory
+     * @param FormFactoryInterface  $formFactory
      */
-    public function __construct(CartProviderInterface $cartProvider, RepositoryInterface $cartItemRepository, FormFactory $formFactory)
+    public function __construct(CartProviderInterface $cartProvider, RepositoryInterface $cartItemRepository, FormFactoryInterface $formFactory)
     {
         $this->cartProvider = $cartProvider;
         $this->cartItemRepository = $cartItemRepository;
@@ -65,9 +65,9 @@ class SyliusCartExtension extends Twig_Extension
     public function getFunctions()
     {
         return array(
-            'sylius_cart_exists'  => new Twig_Function_Method($this, 'hasCart'),
-            'sylius_cart_get'     => new Twig_Function_Method($this, 'getCurrentCart'),
-            'sylius_cart_form'    => new Twig_Function_Method($this, 'getItemFormView'),
+             new \Twig_SimpleFunction('sylius_cart_exists', array($this, 'hasCart')),
+             new \Twig_SimpleFunction('sylius_cart_get', array($this, 'getCurrentCart')),
+             new \Twig_SimpleFunction('sylius_cart_form', array($this, 'getItemFormView')),
         );
     }
 
