@@ -12,17 +12,15 @@
 namespace spec\Sylius\Bundle\MoneyBundle\Twig;
 
 use PhpSpec\ObjectBehavior;
+use Sylius\Bundle\MoneyBundle\Context\CurrencyContextInterface;
+use Sylius\Bundle\MoneyBundle\Converter\CurrencyConverterInterface;
 
 /**
  * @author Paweł Jędrzejewski <pjedrzejewski@diweb.pl>
  */
 class SyliusMoneyExtensionSpec extends ObjectBehavior
 {
-    /**
-     * @param Sylius\Bundle\MoneyBundle\Context\CurrencyContextInterface     $currencyContext
-     * @param Sylius\Bundle\MoneyBundle\Converter\CurrencyConverterInterface $converter
-     */
-    function let($currencyContext, $converter)
+    function let(CurrencyContextInterface $currencyContext, CurrencyConverterInterface $converter)
     {
         $this->beConstructedWith($currencyContext, $converter, 'en');
     }
@@ -47,7 +45,7 @@ class SyliusMoneyExtensionSpec extends ObjectBehavior
         $this->formatMoney(500)->shouldReturn('€5.00');
     }
 
-    function it_allows_to_format_money_in_different_currencies($currencyContext, $converter)
+    function it_allows_to_format_money_in_different_currencies($currencyContext)
     {
         $currencyContext->getDefaultCurrency()->willReturn('EUR');
 
